+++ PanoramaApp/Pages/Movies/MovieListDetails.cshtml.cs
@@ -9,60 +9,40 @@ namespace PanoramaApp.Pages.Movies
     using Microsoft.EntityFrameworkCore;
     using PanoramaApp.Data;
     using PanoramaApp.Models;
+    using PanoramaApp.Interfaces;
 
     public class MovieListDetailsModel : PageModel
     {
-        private readonly ApplicationDbContext context;
+        private readonly IMovieListService _movieListService;
 
-        public MovieListDetailsModel(ApplicationDbContext context)
+        public MovieListDetailsModel(IMovieListService movieListService)
         {
-            this.context = context;
+            _movieListService = movieListService;
         }
 
         public MovieList MovieList { get; set; } = default!;
 
         public async Task<IActionResult> OnGetAsync(int id)
         {
-            this.MovieList = await this.context.MovieLists
-                .Include(ml => ml.Movies)
-                .ThenInclude(mli => mli.Movie)
-                .FirstOrDefaultAsync(ml => ml.Id == id);
+            MovieList = await _movieListService.GetMovieListByIdAsync(id);
 
-            if (this.MovieList == null)
+            if (MovieList == null)
             {
-                return this.RedirectToPage("/Error");
+                return RedirectToPage("/Error");
             }
 
-            return this.Page();
+            return Page();
         }
 
         public async Task<IActionResult> OnPostSavePrioritiesAsync([FromBody] List<MoviePriorityUpdate> updates)
         {
             if (updates == null || !updates.Any())
             {
-                return this.BadRequest("Invalid data received.");
+                return BadRequest("Invalid data received.");
             }
 
-            foreach (var update in updates)
-            {
-                var movie = await this.context.Movies
-                    .FirstOrDefaultAsync(m => m.Id == update.Id);
-
-                if (movie != null)
-                {
-                    movie.Priority = update.Priority;
-                }
-            }
-
-            await this.context.SaveChangesAsync();
+            await _movieListService.UpdateMoviePrioritiesAsync(updates);
             return new JsonResult(new { success = true });
-        }
-
-        public class MoviePriorityUpdate
-        {
-            public int Id { get; set; }
-
-            public int Priority { get; set; }
         }
     }
 }
